--- src/App.js
@@ -1,21 +1,0 @@
-import logo from "./logo.svg";
-import "./App.css";
-
-function App() {
-	return (
-		<div className="App">
-			<header className="App-header">
-				<a
-					className="App-link"
-					href="https://reactjs.org"
-					target="_blank"
-					rel="noopener noreferrer"
-				>
-					Learn React
-				</a>
-			</header>
-		</div>
-	);
-}
-
-export default App;
