+++ src/App.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const App = () => {
+	return (
+		<div className="App">
+			<h1>The website is under construction! Please return later!</h1>
+		</div>
+	);
+};
+
+export default App;
